% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{A line Graph of average population}
\usage{
CountryPopulation(x)
}
\arguments{
\item{x}{A real country that exists in the table WorldPopulation.}
}
\value{
a line graph of the Countries average population from 1950 to 2020.
}
\description{
Because this is a very simple function, my explanation is short. These
paragraphs should explain everything you need to know.
}
\details{
This is still in the description part of the documentation and and it
will be until we see something that indicates a new section.
}
\examples{
function("China") #will give you a line graph of China's average population
}
