% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Cup.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{World Cups total and average attendance in different years.}
\format{
a data frame with 22 observations of 4 variables
\describe{
\item{Date}{The date of observation as a POSIX date format.}
\item{World_Cup}{World Cup in different years, what country, total attendance, average attendance, and number of matches.}
}
}
\source{
\url{https://en.wikipedia.org/wiki/FIFA_World_Cup}
}
\usage{
World_Cup
}
\description{
World Cups total and average attendance in different years.
}
\keyword{datasets}
